requirements.txt
line-bot-sdk
requests
BeautifulSoup4
Flask==2.1.0
gunicorn==20.1.0